requirements.txt
blinker==1.9.0
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
fonttools==4.56.0
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
numpy==2.2.2
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
python-barcode==0.16.1
python-dateutil==2.9.0.post0
reportlab==4.4.7
six==1.17.0
Werkzeug==3.1.3
